[WIP] reorganize files
