+++ timeseries/src/autogluon/timeseries/models/gluonts/abstract_gluonts.py
@@ -6,7 +6,6 @@ from typing import Any, Dict, Iterator, List, Optional, Type
 import gluonts
 import numpy as np
 import pandas as pd
-from gluonts.core.settings import let
 from gluonts.dataset.common import Dataset as GluonTSDataset
 from gluonts.dataset.field_names import FieldName
 from gluonts.model.estimator import Estimator as GluonTSEstimator
@@ -14,7 +13,6 @@ from gluonts.model.forecast import Forecast, QuantileForecast, SampleForecast
 from gluonts.model.predictor import Predictor as GluonTSPredictor
 from pandas.tseries.frequencies import to_offset
 
-from autogluon.common.utils.log_utils import set_logger_verbosity
 from autogluon.core.hpo.constants import RAY_BACKEND
 from autogluon.core.utils import warning_filter
 from autogluon.core.utils.savers import save_pkl
@@ -22,8 +20,6 @@ from autogluon.timeseries.dataset.ts_dataframe import ITEMID, TIMESTAMP, TimeSer
 from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
 from autogluon.timeseries.utils.features import get_categorical_and_continuous_features
 from autogluon.timeseries.utils.warning_filters import disable_root_logger
-
-from .callback import GluonTSEarlyStoppingCallback, TimeLimitCallback
 
 logger = logging.getLogger(__name__)
 gts_logger = logging.getLogger(gluonts.__name__)
@@ -233,41 +229,6 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         else:
             return None
 
-    def _fit(
-        self,
-        train_data: TimeSeriesDataFrame,
-        val_data: Optional[TimeSeriesDataFrame] = None,
-        time_limit: int = None,
-        **kwargs,
-    ) -> None:
-        verbosity = kwargs.get("verbosity", 2)
-        set_logger_verbosity(verbosity, logger=logger)
-        gts_logger.setLevel(logging.ERROR if verbosity <= 3 else logging.INFO)
-
-        if verbosity > 3:
-            logger.warning(
-                "GluonTS logging is turned on during training. Note that losses reported by GluonTS "
-                "may not correspond to those specified via `eval_metric`."
-            )
-
-        self._check_fit_params()
-
-        callbacks = [TimeLimitCallback(time_limit)]
-
-        early_stopping_patience = self._get_model_params().get("early_stopping_patience", None)
-        if early_stopping_patience:
-            callbacks.append(GluonTSEarlyStoppingCallback(early_stopping_patience))
-
-        # update auxiliary parameters
-        self._deferred_init_params_aux(dataset=train_data, callbacks=callbacks, **kwargs)
-
-        estimator = self._get_estimator()
-        with warning_filter(), disable_root_logger(), let(gluonts.env.env, use_tqdm=False):
-            self.gts_predictor = estimator.train(
-                self._to_gluonts_dataset(train_data),
-                validation_data=self._to_gluonts_dataset(val_data),
-            )
-
     def predict(self, data: TimeSeriesDataFrame, quantile_levels: List[float] = None, **kwargs) -> TimeSeriesDataFrame:
         if self.gts_predictor is None:
             raise ValueError("Please fit the model before predicting.")
@@ -279,7 +240,7 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         )
         input_index_type = type(data.index.levels[0][0])
 
-        with warning_filter(), let(gluonts.env.env, use_tqdm=False):
+        with warning_filter(), gluonts.core.settings.let(gluonts.env.env, use_tqdm=False):
             quantiles = quantile_levels or self.quantile_levels
             if not all(0 < q < 1 for q in quantiles):
                 raise ValueError("Invalid quantile value specified. Quantiles must be between 0 and 1 (exclusive).")
