+++ timeseries/tests/unittests/models/test_models.py
@@ -15,7 +15,7 @@ import autogluon.core as ag
 from autogluon.timeseries import TimeSeriesDataFrame, TimeSeriesEvaluator
 from autogluon.timeseries.models import DeepARModel, ETSModel
 from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
-from autogluon.timeseries.models.gluonts import GenericGluonTSModel
+from autogluon.timeseries.models.gluonts import GenericGluonTSMXNetModel
 
 from ..common import DUMMY_TS_DATAFRAME, dict_equal_primitive, get_data_frame_with_item_index
 from .test_autogluon_tabular import TESTABLE_MODELS as TABULAR_TESTABLE_MODELS
@@ -207,7 +207,7 @@ def test_when_fit_called_then_models_train_and_returned_predictor_inference_corr
 
 
 @pytest.mark.parametrize(
-    "model_class", [DeepARModel, ETSModel, partial(GenericGluonTSModel, gluonts_estimator_class=MQRNNEstimator)]
+    "model_class", [DeepARModel, ETSModel, partial(GenericGluonTSMXNetModel, gluonts_estimator_class=MQRNNEstimator)]
 )
 @pytest.mark.parametrize("test_data_index", [["A", "B"], ["C", "D"], ["A"]])
 def test_when_fit_called_then_models_train_and_returned_predictor_inference_aligns_with_time(
@@ -234,7 +234,7 @@ def test_when_fit_called_then_models_train_and_returned_predictor_inference_alig
 
 
 @pytest.mark.parametrize(
-    "model_class", [DeepARModel, ETSModel, partial(GenericGluonTSModel, gluonts_estimator_class=MQRNNEstimator)]
+    "model_class", [DeepARModel, ETSModel, partial(GenericGluonTSMXNetModel, gluonts_estimator_class=MQRNNEstimator)]
 )
 @pytest.mark.parametrize(
     "train_data, test_data",
